Add sign transaction for stellar
