+++ UnstoppableWallet/UnstoppableWallet/Modules/SendNew/WalletConnect/WalletConnectSendHandler.swift
@@ -6,7 +6,7 @@ enum WalletConnectSendHandler {
     static func instance(request: WalletConnectRequest) -> ISendHandler? {
         switch request.payload {
         case is WCEthereumTransactionPayload: return WalletConnectEvmSendHandler.instance(request: request)
-        case is WCStellarTransactionPayload: return WalletConnectStellarSendHandler.instance(request: request)
+        case is WCStellarTransactionPayload: return WalletConnectStellarTransactionHandler.instance(request: request)
         default: return nil
         }
     }
